requirements.txt
python-dotenv==1.0.0
openai==1.12.0
azure-identity==1.14.0
azure-search-documents==11.4.0b8
azure-ai-formrecognizer==3.3.2
streamlit==1.28.0
PyPDF2==3.0.1
Pillow==10.0.1
numpy==1.26.2
autogen==0.2.7